+++ packages/components/barchart/CHANGELOG.md
@@ -1,6 +1,7 @@
-## `@svizzle/barchart` v0.5.1 (next)
+## `@svizzle/barchart` v0.6.0 (next)
 
 - use reference values to build the x scale in order to draw ref lines exceeding the data extent correctly
+- show a customisable message if `items` is not provided or is empty
 
 ## `@svizzle/barchart` v0.5.0